If AudioWorklet AudioParam is constant, make the array have length 1
If the AudioParam is constant over a rendering quantum, we're allowed
to set the parameter value array to have a length one to indicate that
the audio param is constant.

Make this so if the audio param is truly constant.

Update audioworklet-audioparam.https.html test to handle this case.
The automation tests there are actually constant for a render quantum.

Bug: 814796
Test: audioworklet-audioparam.https.html
Change-Id: Ib6e5805dbeb418ad3c202c9c79adc87bc93dad4a